test(storage): run integration suite against real Postgres in CI (#118)
Parameterise the mcs-storage integration suite so the same test bodies run
on both backends. A new `tests::harness` helper, `connect_test_storage`,
reads `MCS_TEST_DATABASE_URL` (defaulting to `sqlite::memory:`) and returns
an isolated, migrated `SqlxStorage`:

- SQLite: a private in-memory database per test (torn down with the pool).
- Postgres: a private, uniquely-named schema per test, with every pool
  connection pinning `search_path` to it via `after_connect`. The schema is
  dropped on teardown. Per-test uniqueness keeps a shared server isolated
  with no cross-test interference.

The old SQLite-only `sqlx_sqlite.rs` becomes the backend-agnostic
`sqlx_backend.rs`, and its `storage()` helper now goes through the harness.

Fix portability bugs the Postgres run surfaced: integer columns the code
binds/reads as `i64` (`ply`, clocks, `rated`) move from `INTEGER` to
`BIGINT`, and Glicko-2 rating columns read as `f64` move from `REAL` to
`DOUBLE PRECISION`. Postgres `INTEGER`/`REAL` are 4-byte and would not
decode as `i64`/`f64`; both widened types keep SQLite's affinity behaviour
unchanged.

Add a `postgres integration tests` CI job with a `postgres:16` service that
sets `MCS_TEST_DATABASE_URL` and runs the suite under
`--no-default-features --features postgres --test-threads=1`. The existing
`ci` job keeps running the suite on SQLite.

+++ crates/mcs-storage/migrations/0003_game_snapshot.sql
@@ -7,9 +7,12 @@
 --
 -- Portability note: as with `0001_init.sql`, this DDL must run unchanged on
 -- both SQLite and PostgreSQL, so it sticks to the lowest common denominator —
--- `ALTER TABLE ... ADD COLUMN` with TEXT/INTEGER types and simple literal
+-- `ALTER TABLE ... ADD COLUMN` with TEXT/BIGINT types and simple literal
 -- defaults. Each `ADD COLUMN` is its own statement (both engines forbid adding
--- several columns in one `ALTER TABLE`).
+-- several columns in one `ALTER TABLE`). The integer columns are BIGINT (not
+-- INTEGER) because the application binds and reads them as 8-byte `i64`, which
+-- matches Postgres `BIGINT`; Postgres `INTEGER` is only 4 bytes and would fail
+-- to decode as `i64`. SQLite stores both with INTEGER affinity regardless.
 
 -- JSON-encoded `mcs_core::VariantOptions`. Existing rows predate per-game
 -- options, so they default to the JSON null literal, which deserialises to
@@ -17,12 +20,12 @@
 ALTER TABLE games ADD COLUMN variant_options TEXT NOT NULL DEFAULT 'null';
 
 -- Live snapshot: half-moves played so far.
-ALTER TABLE games ADD COLUMN ply INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE games ADD COLUMN ply BIGINT NOT NULL DEFAULT 0;
 
 -- Live snapshot: remaining clocks in milliseconds; NULL for untimed games or
 -- before the first snapshot is recorded.
-ALTER TABLE games ADD COLUMN clock_white_ms INTEGER;
-ALTER TABLE games ADD COLUMN clock_black_ms INTEGER;
+ALTER TABLE games ADD COLUMN clock_white_ms BIGINT;
+ALTER TABLE games ADD COLUMN clock_black_ms BIGINT;
 
 -- Live snapshot: whose turn it is, as the lowercase `mcs_core::Color`
 -- discriminant ("white"/"black"); NULL when not applicable.
