+++ crates/mcs-storage/migrations/0005_rated.sql
@@ -6,14 +6,15 @@
 --
 -- Portability note: as with the earlier migrations, this DDL must run unchanged
 -- on both SQLite and PostgreSQL, so it sticks to the lowest common denominator —
--- `ALTER TABLE ... ADD COLUMN` with an INTEGER boolean (0 = casual, 1 = rated),
--- matching how the codebase already stores other small integers. Each
--- `ADD COLUMN` is its own statement (both engines forbid adding several columns
--- in one `ALTER TABLE`).
+-- `ALTER TABLE ... ADD COLUMN` with a BIGINT boolean (0 = casual, 1 = rated),
+-- matching how the codebase already stores other small integers (bound and read
+-- as 8-byte `i64`, which maps onto Postgres `BIGINT`). Each `ADD COLUMN` is its
+-- own statement (both engines forbid adding several columns in one
+-- `ALTER TABLE`).
 --
 -- Existing rows predate the distinction, so they default to `1` (rated),
 -- preserving their original rating behaviour.
 
-ALTER TABLE games ADD COLUMN rated INTEGER NOT NULL DEFAULT 1;
+ALTER TABLE games ADD COLUMN rated BIGINT NOT NULL DEFAULT 1;
 
-ALTER TABLE seeks ADD COLUMN rated INTEGER NOT NULL DEFAULT 1;
+ALTER TABLE seeks ADD COLUMN rated BIGINT NOT NULL DEFAULT 1;
